Basic Message Log Implementation
Added a very basic message log, storing the last 500 messages.

It can be useful in case someone accidentally skipped an important message, wants to check if a certain move has been used or what weather/terrain is currently active.

In addition, as the main menu was a bit full, egg-related options have been moved to a new sub-menu.

To-Do:
- Implement a better UI. (The current one feels a bit slow...)
- Maybe, optionally, also log damage as well as ability and item triggers?
- Currently the log will be purged once the tab is closed/reloaded. I don't think storing it is necessary, but if someone thinks otherwise, it could be implemented. (Probably locally, to reduce storage space usage on the server.)

Also the new text for various languages might require checking...